Promoted several tests from spring-specific tests to also cover jax-rs.
