+++ src/test/java/org/versly/rest/wsdoc/SpringMVCRestAnnotationProcessorTest.java
@@ -87,35 +87,6 @@ public class SpringMVCRestAnnotationProcessorTest extends AbstractRestAnnotation
                 output.contains("multiple-bindings-a<") && output.contains("multiple-bindings-b<"));
     }
 
-    // issue #29
-    @Test
-    public void assertNoRedundantUriParametersForResource() {
-        processResource("RestDocEndpoint.java", "raml", "public");
-        Raml raml = new RamlDocumentBuilder().build(output, "http://example.com");
-        AssertJUnit.assertNotNull("RAML not parseable", raml);
-        Resource resource = raml.getResource("/mount/api/v1/widgets/{id1}/gizmos");
-        AssertJUnit.assertNotNull("Resource /mount/api/v1/widgets/{id1}/gizmos not found", resource);
-        resource = resource.getResource("/{id2}");
-        AssertJUnit.assertNotNull("Resource /mount/api/v1/widgets/{id1}/gizmos/{id2} not found", resource);
-    }
-
-    @Test
-    public void assertUriParameterNormalization() {
-        processResource("UriParameterNormalization.java", "raml", "public");
-        Raml raml = new RamlDocumentBuilder().build(output, "http://example.com");
-        AssertJUnit.assertNotNull("RAML not parseable", raml);
-        Resource resource = raml.getResource("/widgets/{id}");
-        AssertJUnit.assertNotNull("Resource /widgets/{id} not found", resource);
-        UriParameter id = resource.getUriParameters().get("id");
-        AssertJUnit.assertNotNull("Resource /widgets/{id} has no id URI parameter", id);
-        AssertJUnit.assertEquals("Resource /widgets/{id} id URI parameter description is wrong",
-                "The widget identifier.", id.getDescription().trim());
-        resource = resource.getResource("/gadgets");
-        AssertJUnit.assertNotNull("Resource /widgets/{id}/gadgets not found", resource);
-        id = resource.getUriParameters().get("id");
-        AssertJUnit.assertNull("Resource /widgets/{id}/gadgets has it's own id URI parameter when it should not", id);
-    }
-
     @Test
     public void assertAllMethods() {
         super.assertAllMethods();
@@ -125,34 +96,6 @@ public class SpringMVCRestAnnotationProcessorTest extends AbstractRestAnnotation
                     "expected 'allMethodsPatch' in doc string; got: \n" + output,
                     output.contains("allMethodsPatch"));
         }
-    }
-
-    @Test
-    public void testEnumsTypesQueryForRaml() {
-        processResource("RestDocEndpoint.java", "raml", "public");
-        Raml raml = new RamlDocumentBuilder().build(output, "http://example.com");
-        Resource resource = raml.getResource("/mount/api/v1/whirlygigs");
-        AssertJUnit.assertNotNull("Resource /mount/api/v1/whirlygigs not found", resource);
-        Action action = resource.getAction(ActionType.GET);
-        AssertJUnit.assertNotNull("Method GET not found on /mount/api/v1/whirlygigs", action);
-        QueryParameter qp = action.getQueryParameters().get("color");
-        AssertJUnit.assertNotNull("No color query param found on GET method of /mount/api/v1/whirlygigs", qp);
-        List<String> enums = qp.getEnumeration();
-        AssertJUnit.assertNotNull("Color query param on GET method of /mount/api/v1/whirlygigs not enum", enums);
-        AssertJUnit.assertEquals("Color query param on GET /mount/api/v1/whirlygigs is wrong size", 3, enums.size());
-    }
-
-    @Test
-    public void testEnumsTypesInPathForRaml() {
-        processResource("RestDocEndpoint.java", "raml", "public");
-        Raml raml = new RamlDocumentBuilder().build(output, "http://example.com");
-        Resource resource = raml.getResource("/mount/api/v1/colors/{color}");
-        AssertJUnit.assertNotNull("Resource /mount/api/v1/colors/{color} not found", resource);
-        UriParameter up = resource.getUriParameters().get("color");
-        AssertJUnit.assertNotNull("No color path param found on GET method of /mount/api/v1/colors/{color}", up);
-        List<String> enums = up.getEnumeration();
-        AssertJUnit.assertNotNull("Color path param on GET method of /mount/api/v1/colors/{color} not enum", enums);
-        AssertJUnit.assertEquals("Color path param on GET /mount/api/v1/colors/{color} is wrong size", 3, enums.size());
     }
 
     public static void main(String[] args) throws IOException, URISyntaxException {
